Cache storage implementation
